pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "corpusprep"
version = "0.1.0"
description = "Peralatan persiapan corpus regulasi perpajakan Indonesia"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "riset-pajak"}
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pydantic>=2.0",
    "PyYAML>=6.0",
    "pdfplumber>=0.10",
    "python-docx>=1.1",
    "beautifulsoup4>=4.12",
    "httpx>=0.27",
]

[project.scripts]
riset-pajak = "corpusprep.cli:main"

[project.optional-dependencies]
vector = [
    "chromadb>=0.4",
    "sentence-transformers>=2.2",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
corpusprep = ["py.typed"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.pytest.ini_options]
testpaths = ["tests"]
